requirements.txt
numpy
biopython
networkit==5.0
lxml
